chapter-9/Transform-with-AzureAI/src/flows/test_data_generator/requirements.txt
promptflow
requests
python-dotenv
azure-ai-generative[evaluate]==1.0.0b8
promptflow-evals==0.3.0
azure-search-documents
typer
openai>=1.0.0
pandas
rich
jmespath
